Created new rn proyect
There was something wrong with the navigation on the first and second proyect, not sure what was going on, i tried for a couple of days to fix the proyect. up to the point i created a new proyect and installed a navigation from scratch. and it worked, so i went ahead and created a new proyect in the repo with the navigation.